+++ src/fsize.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 #include "fsize.h"
 
-int fsize(FILE *fp)
+long fsize(FILE *fp)
 {
-	int curr = ftell(fp);
+	long curr = ftell(fp);
 	if (fseek(fp, 0, SEEK_END) < 0)
 		return -1;
-	int res = ftell(fp);
+	long res = ftell(fp);
 	if (fseek(fp, curr, SEEK_SET) < 0)
 		return -1;
 	return res;
